# public/api-docs.yaml
openapi: 3.0.0
info:
  title: TheoBot Database API Layer
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    _If you're looking for the Swagger 2.0/OAS 2.0 version of Petstore, then click [here](https://editor.swagger.io/?url=https://petstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load Petstore OAS 2.0` menu option!_
    
    Some useful links:
    - [TelegramBot API Docs](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
servers:
  - url: https://dqpllfj7tb.execute-api.us-east-1.amazonaws.com/dev
paths:
  /projects:
    get:
      tags:
        - Projects
      summary: List all projects
      operationId: listProjects
      responses:
        '200':
          description: An array of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    post:
      tags:
        - Projects
      summary: Create a new project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        '201':
          description: Project created
        '400':
          description: Invalid input
  /projects/{id}:
    get:
      tags:
        - Projects
      summary: Get a project by ID
      operationId: getProject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detailed information about a project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Project not found
    put:
      tags:
        - Projects
      summary: Update a project
      operationId: updateProject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
      responses:
        '200':
          description: Project updated
        '400':
          description: Invalid input
        '404':
          description: Project not found
    delete:
      tags:
        - Projects
      summary: Delete a project
      operationId: deleteProject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Project deleted
        '404':
          description: Project not found
  /projects/byMoon/{moonID}:
    get:
      tags:
        - Projects
      summary: Get projects by moon ID
      operationId: projectsByMoonID
      parameters:
        - name: moonID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: An array of projects associated with a moon ID
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
  /users:
    post:
      tags:
        - Users
      summary: Create a new user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: User created
        '400':
          description: Invalid input
    get:
      tags:
        - Users
      summary: List all users
      operationId: listUsers
      responses:
        '200':
          description: A list of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user by specific id
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
    put:
      tags:
        - Users
      summary: Update a specific user
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User updated
        '400':
          description: Invalid input
        '404':
          description: User not found
    delete:
      tags:
        - Users
      summary: Deletes a user
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User deleted
        '404':
          description: User not found
  /users/verify/{id}:
    post:
      tags:
        - Users
      summary: Verifies a user
      operationId: verifyUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User verified
        '404':
          description: User not found
  /satellites:
   post:
      summary: Create a new satellite
      description: Adds a new satellite to the system
      operationId: createSatellite
      tags:
        - Satellites
      requestBody:
        description: Satellite object that needs to be added
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Satellite'
      responses:
        '201':
          description: Satellite created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Satellite'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /satellites/{id}:
    get:
      tags:
        - Satellites
      summary: Get a satellite by ID
      operationId: getSatellite
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detailed information about a satellite
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Satellite'
        '404':
          description: Satellite not found
    put:
      tags:
        - Satellites
      summary: Update a satellite by ID
      operationId: updateSatellite
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Satellite'
      responses:
        '200':
          description: Satellite updated
        '400':
          description: Invalid input
        '404':
          description: Satellite not found
    delete:
      tags:
        - Satellites
      summary: Delete a satellite by ID
      operationId: deleteSatellite
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Satellite deleted
        '404':
          description: Satellite not found
...
  /files/task/{taskId}:
    get:
      tags:
        - Files
      summary: List all files associated with a specific task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of files
        '500':
          description: Error retrieving task files
  /files/project/{projectId}:
    get:
      tags:
        - Files
      summary: List all files associated with a specific project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of files
        '500':
          description: Error retrieving project files
  /tasks:
    get:
      tags:
        - Tasks
      summary: List all tasks
      operationId: listTasks
      responses:
        '200':
          description: An array of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
    paths:

    post:
      tags:
        - Tasks
      summary: Create a new task
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: The title of the task
                description:
                  type: string
                  description: The description of the task
                assignee:
                  type: string
                  description: The assignee of the task
                projectID:
                  type: string
                  description: The ID of the project the task belongs to
                dueDate:
                  type: string
                  format: date-time
                  description: The due date of the task
                taskType:
                  type: string
                  description: The type of the task
                notes:
                  type: string
                  description: Any notes related to the task
                status:
                  type: string
                  description: The status of the task
      responses:
        '200':
          description: Task successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '500':
          description: Error creating task
  /tasks/{id}:
    get:
      tags:
        - Tasks
      summary: Get a task by ID
      operationId: getTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detailed information about a task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found
    put:
      tags:
        - Tasks
      summary: Update a task
      operationId: updateTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '200':
          description: Task updated
        '400':
          description: Invalid input
        '404':
          description: Task not found
    delete:
      tags:
        - Tasks
      summary: Delete a task
      operationId: deleteTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task deleted
        '404':
          description: Task not found
  /tasks/byProject/{projectID}:
    get:
      tags:
        - Tasks
      summary: Get tasks by project ID
      operationId: tasksByProjectID
      parameters:
        - name: projectID
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: An array of tasks associated with a project ID
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
...
components:
  schemas:
    Satellite:
      type: object
      required:
        - id
        - moonID
        - title
      properties:
        id:
          type: string
          description: Unique identifier for the Satellite
        moonID:
          type: string
          description: Identifier for the Moon that the Satellite is associated with
        title:
          type: string
          description: Title of the Satellite
        description:
          type: string
          description: Description of the Satellite
        User:
          $ref: '#/components/schemas/User'
    User:
      type: object
      required:
        - id
        - firstName
        - username
        - completedOnboarding
        - role
        - status
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        username:
          type: string
        completedOnboarding:
          type: boolean
        role:
          $ref: '#/components/schemas/UserRole'
        bio:
          type: string
        status:
          $ref: '#/components/schemas/UserStatus'
    UserRole:
      type: string
      enum:
        - SUPER_ADMIN
        - ADMIN
        - MEMBER
        - VIEWER
    UserStatus:
      type: string
      enum:
        - PENDING
        - CONFIRMED
        - SUSPENDED
...
    Project:
      type: object
      required:
        - id
        - moonID
        - name
        - status
        - projectLead
      properties:
        id:
          type: string
        moonID:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum:
            - ACTIVE
            - COMPLETED
            - ON_HOLD
        projectLead:
          type: string
          description: Name of the project lead. Used for Jira project creation.
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        members:
          type: array
          items:
            $ref: '#/components/schemas/User'
        roles:
          type: array
          items:
            $ref: '#/components/schemas/ProjectRole'
        files:
          type: array
          items:
            $ref: '#/components/schemas/File'
        chats:
          type: array
          items:
            $ref: '#/components/schemas/Chat'
...
    File:
      type: object
      required:
        - id
        - name
        - url
      properties:
        id:
          type: string
        name:
          type: string
        url:
          type: string
        uploadedBy:
          $ref: '#/components/schemas/User'
    Task:
      type: object
      required:
        - id
        - projectID
        - title
        - status
      properties:
        id:
          type: string
        projectID:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/TaskStatus'
        assignedUsers:
          type: array
          items:
            $ref: '#/components/schemas/User'
        dueDate:
          type: string
          format: date
        files:
          type: array
          items:
            $ref: '#/components/schemas/File'
        chats:
          type: array
          items:
            $ref: '#/components/schemas/Chat'
        notes:
          type: array
          items:
            type: string
    TaskStatus:
      type: string
      enum:
        - TO_DO
        - IN_PROGRESS
        - DONE
    Chat:
      type: object
      required:
        - id
        - message
        - sentAt
        - sentBy
      properties:
        id:
          type: string
        message:
          type: string
        sentAt:
          type: string
          format: date-time
        sentBy:
          $ref: '#/components/schemas/User'
...
    ProjectRole:
      type: object
      required:
        - id
        - projectID
        - title
      properties:
        id:
          type: string
        projectID:
          type: string
          description: The ID of the project this role is associated with.
        title:
          type: string
          description: The title of the project role.
        description:
          type: string
          description: A brief description of the role.
        assignedTo:
          $ref: '#/components/schemas/User'
          description: The user who is assigned to this role.